requirements.txt
asgiref==3.8.1
Django==5.1.3
pillow==11.0.0
sqlparse==0.5.1
tzdata==2024.2
dj-database-url==2.2.0
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
waitress==3.0.0
whitenoise==6.7.0
python-dotenv==1.0.1
